refactor: Restructure repository and clean up a lot.
