+++ src/util/telegram.js
@@ -1,8 +1,10 @@
+const { flaschenpost } = require("flaschenpost");
 const R = require("ramda");
 
-const { getters } = require("../leetbot/store/getters");
+const { getters } = require("../store/getters");
 
 const { languageOrDefault } = getters;
+const logger = flaschenpost.getLogger();
 
 const chatIdInContext = R.path(["chat", "id"]);
 
@@ -10,6 +12,10 @@ const fromIdInContext = R.path(["from", "id"]);
 
 const messageIdInContext = R.path(["update", "message", "message_id"]);
 
+/**
+ * Get the name of the user in the given context. Tries various ways to retrieve
+ * the name to get something legible.
+ */
 const legibleUserInContext = R.compose(
   R.head,
   R.dropWhile(R.isNil),
@@ -17,36 +23,46 @@ const legibleUserInContext = R.compose(
   R.prop("from")
 );
 
+/**
+ * Retrieve the text message in the given context.
+ */
 const messageInContext = R.path(["update", "message", "text"]);
 
+/**
+ * Parse the context's message of the form
+ *
+ * /command subcommand some more stuff
+ *
+ * to retrieve the subcommand.
+ */
 const subCommandInContext = ctx => {
   const message = messageInContext(ctx);
   if (message[0] !== "/") {
     return undefined;
   }
-  return R.pipe(
-    R.split(" "),
-    R.tail,
-    R.join(" ")
-  )(message);
+  return R.pipe(R.split(" "), R.tail, R.join(" "))(message);
 };
 
 /**
  * Catch and log any error that might occur further down the line.
- *
- * @param {*} ctx
- * @param {*} next
  */
 const crashHandler = (ctx, next) => {
-  return next().catch(console.log);
+  return next().catch(error => logger.error("An error occured.", { error }));
 };
 
+/**
+ * Retrieves the configured language for the chat in context (or uses the i18n
+ * default language) and attaches a translation function to the context for easy
+ * access in later middlewares.
+ */
 const translationMiddleware = ({ i18n, store }) => (ctx, next) => {
   const chatId = chatIdInContext(ctx);
+
   ctx.t = (key, params) => {
     const language = languageOrDefault(chatId, store);
     return i18n.t(key, { ...params, lng: language });
   };
+
   return next();
 };
